TelegramBot/pyproject.toml
[project]
name = "telegram-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "ycalk",email = "alosev752@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiogram (>=3.22.0,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "telegram_bot", from = "src"}]

[tool.poetry.scripts]
bot = "telegram_bot:run"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
